refactor dict converters
