+++ database/event_query.py
@@ -2,7 +2,7 @@ from google.appengine.ext import ndb
 
 from consts.district_type import DistrictType
 from database.database_query import DatabaseQuery
-from helpers.model_to_dict import ModelToDict
+from database.dict_converters.event_converter import EventConverter
 from models.event import Event
 from models.event_team import EventTeam
 from models.team import Team
@@ -13,7 +13,7 @@ class EventListQuery(DatabaseQuery):
     CACHE_KEY_FORMAT = 'event_list_{}'  # (year)
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         year = self._query_args[0]
         events = yield Event.query(Event.year == year).fetch_async()
         raise ndb.Return(events)
@@ -24,7 +24,7 @@ class DistrictEventsQuery(DatabaseQuery):
     CACHE_KEY_FORMAT = 'district_events_{}'  # (district_key)
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         district_key = self._query_args[0]
         year = int(district_key[:4])
         district_abbrev = district_key[4:]
@@ -38,24 +38,24 @@ class DistrictEventsQuery(DatabaseQuery):
 class TeamEventsQuery(DatabaseQuery):
     CACHE_VERSION = 1
     CACHE_KEY_FORMAT = 'team_events_{}'  # (team_key)
+    DICT_CONVERTER = EventConverter
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         team_key = self._query_args[0]
         event_teams = yield EventTeam.query(EventTeam.team == ndb.Key(Team, team_key)).fetch_async()
         event_keys = map(lambda event_team: event_team.event, event_teams)
         events = yield ndb.get_multi_async(event_keys)
-        if dict_version:
-            events = ModelToDict.convertEvents(events, dict_version)
         raise ndb.Return(events)
 
 
 class TeamYearEventsQuery(DatabaseQuery):
     CACHE_VERSION = 1
     CACHE_KEY_FORMAT = 'team_year_events_{}_{}'  # (team_key, year)
+    DICT_CONVERTER = EventConverter
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         team_key = self._query_args[0]
         year = self._query_args[1]
         event_teams = yield EventTeam.query(
@@ -63,6 +63,4 @@ class TeamYearEventsQuery(DatabaseQuery):
             EventTeam.year == year).fetch_async()
         event_keys = map(lambda event_team: event_team.event, event_teams)
         events = yield ndb.get_multi_async(event_keys)
-        if dict_version:
-            events = ModelToDict.convertEvents(events, dict_version)
         raise ndb.Return(events)
