+++ database/match_query.py
@@ -1,6 +1,7 @@
 from google.appengine.ext import ndb
 
 from database.database_query import DatabaseQuery
+from database.dict_converters.match_converter import MatchConverter
 from helpers.model_to_dict import ModelToDict
 from models.event import Event
 from models.match import Match
@@ -11,7 +12,7 @@ class EventMatchesQuery(DatabaseQuery):
     CACHE_KEY_FORMAT = 'event_matches_{}'  # (event_key)
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         event_key = self._query_args[0]
         match_keys = yield Match.query(Match.event == ndb.Key(Event, event_key)).fetch_async(keys_only=True)
         matches = yield ndb.get_multi_async(match_keys)
@@ -21,32 +22,30 @@ class EventMatchesQuery(DatabaseQuery):
 class TeamEventMatchesQuery(DatabaseQuery):
     CACHE_VERSION = 0
     CACHE_KEY_FORMAT = 'team_event_matches_{}_{}'  # (team_key, event_key)
+    DICT_CONVERTER = MatchConverter
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         team_key = self._query_args[0]
         event_key = self._query_args[1]
         match_keys = yield Match.query(
             Match.team_key_names == team_key,
             Match.event == ndb.Key(Event, event_key)).fetch_async(keys_only=True)
         matches = yield ndb.get_multi_async(match_keys)
-        if dict_version:
-            matches = ModelToDict.convertMatches(matches, dict_version)
         raise ndb.Return(matches)
 
 
 class TeamYearMatchesQuery(DatabaseQuery):
     CACHE_VERSION = 0
     CACHE_KEY_FORMAT = 'team_year_matches_{}_{}'  # (team_key, year)
+    DICT_CONVERTER = MatchConverter
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         team_key = self._query_args[0]
         year = self._query_args[1]
         match_keys = yield Match.query(
             Match.team_key_names == team_key,
             Match.year == year).fetch_async(keys_only=True)
         matches = yield ndb.get_multi_async(match_keys)
-        if dict_version:
-            matches = ModelToDict.convertMatches(matches, dict_version)
         raise ndb.Return(matches)
