+++ database/team_query.py
@@ -1,8 +1,9 @@
 from google.appengine.ext import ndb
 
 from consts.district_type import DistrictType
+from database.dict_converters.district_converter import DistrictConverter
+from database.dict_converters.team_converter import TeamConverter
 from database.database_query import DatabaseQuery
-from helpers.model_to_dict import ModelToDict
 from models.district_team import DistrictTeam
 from models.event import Event
 from models.event_team import EventTeam
@@ -12,13 +13,12 @@ from models.team import Team
 class TeamQuery(DatabaseQuery):
     CACHE_VERSION = 0
     CACHE_KEY_FORMAT = 'team_{}'  # (team_key)
+    DICT_CONVERTER = TeamConverter
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         team_key = self._query_args[0]
         team = yield Team.get_by_id_async(team_key)
-        if dict_version:
-            team = ModelToDict.convertTeams([team], dict_version)[0]
         raise ndb.Return(team)
 
 
@@ -26,24 +26,24 @@ class TeamListQuery(DatabaseQuery):
     CACHE_VERSION = 1
     CACHE_KEY_FORMAT = 'team_list_{}'  # (page_num)
     PAGE_SIZE = 500
+    DICT_CONVERTER = TeamConverter
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         page_num = self._query_args[0]
         start = self.PAGE_SIZE * page_num
         end = start + self.PAGE_SIZE
         teams = yield Team.query(Team.team_number >= start, Team.team_number < end).fetch_async()
-        if dict_version:
-            teams = ModelToDict.convertTeams(teams, dict_version)
         raise ndb.Return(teams)
 
 
 class TeamListYearQuery(DatabaseQuery):
     CACHE_VERSION = 1
     CACHE_KEY_FORMAT = 'team_list_year_{}_{}'  # (year, page_num)
+    DICT_CONVERTER = TeamConverter
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         year = self._query_args[0]
         page_num = self._query_args[1]
 
@@ -56,8 +56,6 @@ class TeamListYearQuery(DatabaseQuery):
             year_team_keys.add(team_key)
 
         teams = filter(lambda team: team.key.id() in year_team_keys, teams_future.get_result())
-        if dict_version:
-            teams = ModelToDict.convertTeams(teams, dict_version)
         raise ndb.Return(teams)
 
 
@@ -66,7 +64,7 @@ class DistrictTeamsQuery(DatabaseQuery):
     CACHE_KEY_FORMAT = 'district_teams_{}'  # (district_key)
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         district_key = self._query_args[0]
         year = int(district_key[:4])
         district_abbrev = district_key[4:]
@@ -84,7 +82,7 @@ class EventTeamsQuery(DatabaseQuery):
     CACHE_KEY_FORMAT = 'event_teams_{}'  # (event_key)
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         event_key = self._query_args[0]
         event_teams = yield EventTeam.query(EventTeam.event == ndb.Key(Event, event_key)).fetch_async()
         team_keys = map(lambda event_team: event_team.team, event_teams)
@@ -97,7 +95,7 @@ class TeamParticipationQuery(DatabaseQuery):
     CACHE_KEY_FORMAT = 'team_participation_{}'  # (team_key)
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         team_key = self._query_args[0]
         event_teams = yield EventTeam.query(EventTeam.team == ndb.Key(Team, team_key)).fetch_async(keys_only=True)
         years = map(lambda event_team: int(event_team.id()[:4]), event_teams)
@@ -107,20 +105,15 @@ class TeamParticipationQuery(DatabaseQuery):
 class TeamDistrictsQuery(DatabaseQuery):
     CACHE_VERSION = 0
     CACHE_KEY_FORMAT = 'team_districts_{}'  # (team_key)
+    DICT_CONVERTER = DistrictConverter
 
     @ndb.tasklet
-    def _query_async(self, dict_version):
+    def _query_async(self):
         team_key = self._query_args[0]
         district_team_keys = yield DistrictTeam.query(DistrictTeam.team == ndb.Key(Team, team_key)).fetch_async(keys_only=True)
-        if dict_version:
-            ret = ModelToDict.convertDistricts(district_team_keys, dict_version)
-        else:
-            # This conversion really shouldn't happen and the raw district_team_keys should be stored
-            # -fangeugene 2017/01/04
-            ret = {}
-            for district_team_key in district_team_keys:
-                district_key = district_team_key.id().split('_')[0]
-                year = int(district_key[:4])
-                ret[year] = district_key
-
+        ret = {}
+        for district_team_key in district_team_keys:
+            district_key = district_team_key.id().split('_')[0]
+            year = int(district_key[:4])
+            ret[year] = district_key
         raise ndb.Return(ret)
